feat: add user

+++ src/main/java/com/example/demo/mapper/UserMapper.java
@@ -6,7 +6,7 @@ import com.example.demo.entity.User;
 public class UserMapper {
     public  static UserDto toUserDto (User user){
         UserDto tmp = new UserDto();
-        tmp.setId(user.getId());
+        tmp.setId((int) user.getId());
         tmp.setName(user.getName());
         tmp.setEmail(user.getEmail());
         tmp.setPhone(user.getPhone());
